+++ Stitch/Graph/Node/View/NodeTag/NodeTagMenuButtonsView.swift
@@ -40,7 +40,6 @@ struct NodeTagMenuButtonsView: View {
      
     @MainActor
     var moreThanOneNodeSelected: Bool {
-//        graph.selectedNodeIds.count > 1
         graph.selectedCanvasItems.count > 1
     }
